refactor: move lib/shared/child-writer to lib/child-writer
The `child-writer` module is not re-used by both the GUI and the CLI, so
it makes sense to have it in `lib/child-writer`.
