+++ lib/child-writer/cli.js
