+++ lib/child-writer/constants.js
@@ -29,7 +29,7 @@ module.exports = {
    * @property {String} PROJECT_ROOT
    * @memberof CONSTANTS
    */
-  PROJECT_ROOT: path.join(__dirname, '..', '..', '..'),
+  PROJECT_ROOT: path.join(__dirname, '..', '..'),
 
   /**
    * @property {String} WRITER_PROXY_SCRIPT
