+++ lib/child-writer/index.js
@@ -23,8 +23,8 @@ const ipc = require('node-ipc');
 const rendererUtils = require('./renderer-utils');
 const cli = require('./cli');
 const CONSTANTS = require('./constants');
-const EXIT_CODES = require('../exit-codes');
-const robot = require('../robot');
+const EXIT_CODES = require('../shared/exit-codes');
+const robot = require('../shared/robot');
 
 /**
  * @summary Perform a write
