+++ lib/child-writer/renderer-utils.js
