+++ lib/child-writer/utils.js
