+++ lib/child-writer/writer-proxy.js
@@ -25,8 +25,8 @@ const os = require('os');
 const path = require('path');
 const sudoPrompt = Bluebird.promisifyAll(require('sudo-prompt'));
 const utils = require('./utils');
-const EXIT_CODES = require('../exit-codes');
-const packageJSON = require('../../../package.json');
+const EXIT_CODES = require('../shared/exit-codes');
+const packageJSON = require('../../package.json');
 
 // This script is in charge of spawning the writer process and
 // ensuring it has the necessary privileges. It might look a bit
