+++ lib/gui/modules/image-writer.js
@@ -21,7 +21,7 @@
  */
 
 const angular = require('angular');
-const childWriter = require('../../shared/child-writer');
+const childWriter = require('../../child-writer');
 
 const MODULE_NAME = 'Etcher.Modules.ImageWriter';
 const imageWriter = angular.module(MODULE_NAME, [
